chore: move lsp channel creation code out of api
