+++ service/models.go
@@ -1,0 +1,9 @@
+package service
+
+import (
+	"github.com/getAlby/nostr-wallet-connect/models/lnclient"
+)
+
+type Service interface {
+	GetLNClient() lnclient.LNClient
+}
